% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_catch_info}
\alias{append_catch_info}
\title{Append Stock Synthesis 3.30 Catch Info}
\usage{
append_catch_info(
  x = NULL,
  fleet_type = 0,
  timing = 0,
  area = 0,
  units = 0,
  use_catch_mult = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{fleet_type}{\code{\link[=integer]{integer()}}}

\item{timing}{\code{\link[=numeric]{numeric()}}}

\item{area}{\code{\link[=integer]{integer()}}}

\item{units}{\code{\link[=integer]{integer()}}}

\item{use_catch_mult}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_catch_info]{create_catch_info()}} (see examples).
}
\examples{
\dontrun{
create_catch_info() |>
  append_catch_info(1, -1, 1, 1, 1, name = "Fleet name")
}

}
