% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_catchability_info}
\alias{append_catchability_info}
\title{Append Stock Synthesis 3.30 Catchability Info}
\usage{
append_catchability_info(
  x = NULL,
  fleet = 0,
  link_type = 0,
  link_info = 0,
  extra_se = 0,
  bias_adjust = 0,
  float = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{fleet}{\code{\link[=integer]{integer()}}}

\item{link_type}{\code{\link[=integer]{integer()}}}

\item{link_info}{\code{\link[=integer]{integer()}}}

\item{extra_se}{\code{\link[=numeric]{numeric()}}}

\item{bias_adjust}{\code{\link[=numeric]{numeric()}}}

\item{float}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_catchability_info]{create_catchability_info()}} (see examples).
}
\examples{
\dontrun{
create_catchability_info() |>
  append_catchability_info(1, 1, 0, 0, 0, 1,  name = "Fleet name")
}

}
