% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_index_info}
\alias{append_index_info}
\title{Append Stock Synthesis 3.30 Index Info}
\usage{
append_index_info(
  x = NULL,
  fleet = 0,
  units = 0,
  error_type = 0,
  sd_report = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{fleet}{\code{\link[=integer]{integer()}}}

\item{units}{\code{\link[=integer]{integer()}}}

\item{error_type}{\code{\link[=numeric]{numeric()}}}

\item{sd_report}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_index_info]{create_index_info()}} (see examples).
}
\examples{
\dontrun{
create_index_info() |>
  append_index_info(1, 1, 0, 0, name = "Fleet name")
}

}
