% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_lambda_info}
\alias{append_lambda_info}
\title{Append Stock Synthesis 3.30 Lambda Info}
\usage{
append_lambda_info(
  x = NULL,
  component = 0,
  fleet = 0,
  phase = 0,
  value = 0,
  method = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{component}{\code{\link[=integer]{integer()}}}

\item{fleet}{\code{\link[=integer]{integer()}}}

\item{phase}{\code{\link[=integer]{integer()}}}

\item{value}{\code{\link[=numeric]{numeric()}}}

\item{method}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_lambda_info]{create_lambda_info()}} (see examples).
}
\examples{
\dontrun{
create_lambda_info() |>
  append_lambda_info(1, 1, 1, 1, 1, name = "Fleet name")
}

}
