% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_length_info}
\alias{append_length_info}
\title{Append Stock Synthesis 3.30 Length Info}
\usage{
append_length_info(x = NULL, ..., name = c(""))
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{...}{\code{\link[=numeric]{numeric()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_length_info]{create_length_info()}} (see examples).
}
\examples{
\dontrun{
create_length_info() |>
  append_length_info(-1, 1e-04, 0, 0, 0, 0, 0.001, name = "Fleet name")
}

}
