% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_recruitment_info}
\alias{append_recruitment_info}
\title{Append Stock Synthesis 3.30 Recruitment Info}
\usage{
append_recruitment_info(
  x = NULL,
  growth_pattern = 0,
  month = 0,
  area = 0,
  age_at_settlement = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{growth_pattern}{\code{\link[=integer]{integer()}}}

\item{month}{\code{\link[=integer]{integer()}}}

\item{area}{\code{\link[=integer]{integer()}}}

\item{age_at_settlement}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_recruitment_info]{create_recruitment_info()}} (see examples).
}
\examples{
\dontrun{
create_recruitment_info() |>
  append_recruitment_info(1, 1, 1, 0, name = "Morph 1")
}

}
