% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_seasonality_info}
\alias{append_seasonality_info}
\title{Append Stock Synthesis 3.30 Seasonality Info}
\usage{
append_seasonality_info(
  x = NULL,
  wt_len_fem_1 = 0,
  wt_len_fem_2 = 0,
  maturity_1 = 0,
  maturity_2 = 0,
  fecundity_1 = 0,
  fecundity_2 = 0,
  wt_len_mal_1 = 0,
  wt_len_mal_2 = 0,
  l1 = 0,
  von_bert_k = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{wt_len_fem_1}{\code{\link[=integer]{integer()}}}

\item{wt_len_fem_2}{\code{\link[=integer]{integer()}}}

\item{maturity_1}{\code{\link[=integer]{integer()}}}

\item{maturity_2}{\code{\link[=integer]{integer()}}}

\item{fecundity_1}{\code{\link[=integer]{integer()}}}

\item{fecundity_2}{\code{\link[=integer]{integer()}}}

\item{wt_len_mal_1}{\code{\link[=integer]{integer()}}}

\item{wt_len_mal_2}{\code{\link[=integer]{integer()}}}

\item{l1}{\code{\link[=integer]{integer()}}}

\item{von_bert_k}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_seasonality_info]{create_seasonality_info()}} (see examples).
}
\examples{
\dontrun{
create_seasonality_info() |>
  append_seasonality_info(name = "seasonality")
}

}
