% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append-info.R
\name{append_selectivity_info}
\alias{append_selectivity_info}
\title{Append Stock Synthesis 3.30 Selectivity Info}
\usage{
append_selectivity_info(
  x = NULL,
  pattern = 0,
  discard = 0,
  male = 0,
  special = 0,
  name = c("")
)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{pattern}{\code{\link[=integer]{integer()}}}

\item{discard}{\code{\link[=integer]{integer()}}}

\item{male}{\code{\link[=integer]{integer()}}}

\item{special}{\code{\link[=integer]{integer()}}}

\item{name}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Append an info row to an info \code{\link[tibble:tibble]{tibble::tibble()}} created by
\code{\link[=create_selectivity_info]{create_selectivity_info()}} (see examples).
}
\examples{
\dontrun{
create_selectivity_info() |>
  append_selectivity_info(24, 0, 0, 0, name = "Fleet name")
}

}
