% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_catch_info}
\alias{create_catch_info}
\title{Create Stock Synthesis 3.30 Catch Info Data Frame}
\usage{
create_catch_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_catch_info]{append_catch_info()}} and the pipe operator (see examples).
}
\examples{
\dontrun{
create_catch_info() |>
  append_catch_info(1, -1, 1, 1, 1, name = "Fleet name")
}

}
