% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_catchability_info}
\alias{create_catchability_info}
\title{Create Stock Synthesis 3.30 Catchability Info Data Frame}
\usage{
create_catchability_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_catchability_info]{append_catchability_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_catchability_info() |>
  append_catchability_info(1, 1, 0, 0, 0, 1,  name = "Fleet name")
}

}
