% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_index_info}
\alias{create_index_info}
\title{Create Stock Synthesis 3.30 Index Info Data Frame}
\usage{
create_index_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_index_info]{append_index_info()}} and the pipe operator (see examples).
}
\examples{
\dontrun{
create_index_info() |>
  append_index_info(1, 1, 0, 0, name = "Fleet name")
}

}
