% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_lambda_info}
\alias{create_lambda_info}
\title{Create Stock Synthesis 3.30 Lambda Info Data Frame}
\usage{
create_lambda_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_lambda_info]{append_lambda_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_lambda_info() |>
  append_lambda_info(1, 1, 1, 1, 1, name = "Fleet name")
}

}
