% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_length_info}
\alias{create_length_info}
\title{Create Stock Synthesis 3.30 Length Info Data Frame}
\usage{
create_length_info(use_length_composition = 1L)
}
\arguments{
\item{use_length_composition}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_length_info]{append_length_info()}} and the pipe operator (see examples).
}
\examples{
\dontrun{
create_length_info() |>
  append_length_info(-1, 1e-04, 0, 0, 0, 0, 0.001, name = "Fleet name")
}

}
