% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-p.R
\name{create_p}
\alias{create_p}
\title{Create Stock Synthesis 3.30 Parameter Data Frame}
\usage{
create_p()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create Stock Synthesis 3.30 Parameter Data Frame
}
\details{
Create a parameter \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append parameter
rows using \code{\link[=append_p]{append_p()}} and the pipe operator (see examples).
}
\examples{
\dontrun{
create_p() |>
  append_p(5.0, 15, 9.5, 0, 0, 0, 1, name = "SR_LN(R0)")
}

}
