% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_recruitment_info}
\alias{create_recruitment_info}
\title{Create Stock Synthesis 3.30 Recruitment Info Data Frame}
\usage{
create_recruitment_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_recruitment_info]{append_recruitment_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_recruitment_info() |>
  append_recruitment_info(1, 1, 1, 0, name = "Morph 1")
}

}
