% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_seasonality_info}
\alias{create_seasonality_info}
\title{Create Stock Synthesis 3.30 Seasonality Info Data Frame}
\usage{
create_seasonality_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_seasonality_info]{append_seasonality_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_seasonality_info() |>
  append_seasonality_info(name = "seasonality")
}

}
