% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_selectivity_info}
\alias{create_selectivity_info}
\title{Create Stock Synthesis 3.30 Selectivity Info Data Frame}
\usage{
create_selectivity_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_selectivity_info]{append_selectivity_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_selectivity_info() |>
  append_selectivity_info(24, 0, 0, 0, name = "Fleet name")
}

}
