% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-info.R
\name{create_variance_info}
\alias{create_variance_info}
\title{Create Stock Synthesis 3.30 Variance Info Data Frame}
\usage{
create_variance_info()
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Create an info \code{\link[tibble:tibble]{tibble::tibble()}} with 0 rows. Append info
rows using \code{\link[=append_variance_info]{append_variance_info()}} and the pipe operator
(see examples).
}
\examples{
\dontrun{
create_variance_info() |>
  append_variance_info(4, 1, 1, name = "Fleet name")
}

}
