% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-argument-line.R
\name{write_argument_line}
\alias{write_argument_line}
\title{Write Stock Synthesis 3.30 Argument Line}
\usage{
write_argument_line(x, file, append = TRUE)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{file}{\code{\link[=character]{character()}}}

\item{append}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=character]{character()}}
}
\description{
Write Stock Synthesis 3.30 Argument Line
}
\examples{
\dontrun{
some_descriptive_name <- rep(0, 10)
write_argument_line(x = some_descriptive_name, file = "temp.ss")
}

}
