% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-comment-line.R
\name{write_comment_line}
\alias{write_comment_line}
\title{Write Stock Synthesis 3.30 Comment Line}
\usage{
write_comment_line(x, file, append = TRUE)
}
\arguments{
\item{x}{\code{\link[=character]{character()}}}

\item{file}{\code{\link[=character]{character()}}}

\item{append}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=character]{character()}}
}
\description{
Write Stock Synthesis 3.30 Comment Line
}
\examples{
\dontrun{
write_comment_line("A new comment", file = "temp.ss")
}

}
