% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-control.R
\name{write_control}
\alias{write_control}
\title{Write Stock Synthesis 3.30 Input Control File}
\usage{
write_control(
  file = "control.ss",
  use_weight_at_age = 0,
  n_morphs = 1,
  n_platoons_per_morph = 1,
  recruitment_method = 4,
  recruitment_spatial = 1,
  n_recruitment_settle = 1,
  recruitment_unused = 0,
  recruitment_info,
  n_block_patterns = 0,
  time_varying_method = 0,
  time_varying_auto_generation = rep(0, 5),
  natural_mortality_option = 0,
  growth_method = 1,
  growth_age_at_l1,
  growth_age_at_l2,
  growth_exp_decay,
  growth_feature = 0,
  growth_add_sd = 0,
  growth_cv_option = 0,
  maturity_option,
  maturity_data,
  maturity_first_age,
  fecundity_option,
  hermaphroditism_option = 0,
  parameter_offset_method,
  mortality_growth_parameters,
  seasonal_parameters,
  spawner_recruitment_option,
  spawner_recruitment_parameters,
  use_steepness,
  spawner_recruitment_feature = 0,
  recruitment_deviation_option = 0,
  recruitment_deviation_year_start,
  recruitment_deviation_year_end,
  recruitment_deviation_phase,
  recruitment_deviation_advanced = 0,
  fishing_method,
  fishing_ballpark,
  fishing_ballpark_year,
  fishing_maximum,
  fishing_iterations,
  catchability_info,
  catchability_parameters,
  selectivity_size_info,
  selectivity_age_info,
  selectivity_parameters,
  use_tag_recapture = 0,
  use_selectivity_2d = 0,
  variance_info,
  lambda_max_phase,
  lambda_sd_offset,
  lambda_info,
  sd_report_option = 0
)
}
\arguments{
\item{file}{\code{\link[=character]{character()}}}

\item{use_weight_at_age}{\code{\link[=integer]{integer()}}}

\item{n_morphs}{\code{\link[=integer]{integer()}}}

\item{n_platoons_per_morph}{\code{\link[=integer]{integer()}}}

\item{recruitment_method}{\code{\link[=integer]{integer()}}}

\item{recruitment_spatial}{\code{\link[=integer]{integer()}}}

\item{n_recruitment_settle}{\code{\link[=integer]{integer()}}}

\item{recruitment_unused}{\code{\link[=integer]{integer()}}}

\item{recruitment_info}{\code{\link[=data.frame]{data.frame()}}}

\item{n_block_patterns}{\code{\link[=integer]{integer()}}}

\item{time_varying_method}{\code{\link[=integer]{integer()}}}

\item{time_varying_auto_generation}{\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}}}

\item{natural_mortality_option}{\code{\link[=integer]{integer()}}}

\item{growth_method}{\code{\link[=integer]{integer()}}}

\item{growth_age_at_l1}{\code{\link[=numeric]{numeric()}}}

\item{growth_age_at_l2}{\code{\link[=numeric]{numeric()}}}

\item{growth_exp_decay}{\code{\link[=numeric]{numeric()}}}

\item{growth_feature}{\code{\link[=integer]{integer()}}}

\item{growth_add_sd}{\code{\link[=numeric]{numeric()}}}

\item{growth_cv_option}{\code{\link[=integer]{integer()}}}

\item{maturity_option}{\code{\link[=integer]{integer()}}}

\item{maturity_data}{\code{\link[=data.frame]{data.frame()}}}

\item{maturity_first_age}{\code{\link[=numeric]{numeric()}}}

\item{fecundity_option}{\code{\link[=integer]{integer()}}}

\item{hermaphroditism_option}{\code{\link[=integer]{integer()}}}

\item{parameter_offset_method}{\code{\link[=integer]{integer()}}}

\item{mortality_growth_parameters}{\code{\link[=data.frame]{data.frame()}}}

\item{seasonal_parameters}{\code{\link[=data.frame]{data.frame()}}}

\item{spawner_recruitment_option}{\code{\link[=integer]{integer()}}}

\item{spawner_recruitment_parameters}{\code{\link[=data.frame]{data.frame()}}}

\item{use_steepness}{\code{\link[=integer]{integer()}}}

\item{spawner_recruitment_feature}{\code{\link[=integer]{integer()}}}

\item{recruitment_deviation_option}{\code{\link[=integer]{integer()}}}

\item{recruitment_deviation_year_start}{\code{\link[=integer]{integer()}}}

\item{recruitment_deviation_year_end}{\code{\link[=integer]{integer()}}}

\item{recruitment_deviation_phase}{\code{\link[=integer]{integer()}}}

\item{recruitment_deviation_advanced}{\code{\link[=integer]{integer()}}}

\item{fishing_method}{\code{\link[=integer]{integer()}}}

\item{fishing_ballpark}{\code{\link[=numeric]{numeric()}}}

\item{fishing_ballpark_year}{\code{\link[=integer]{integer()}}}

\item{fishing_maximum}{\code{\link[=numeric]{numeric()}}}

\item{fishing_iterations}{\code{\link[=integer]{integer()}}}

\item{catchability_info}{\code{\link[=data.frame]{data.frame()}}}

\item{catchability_parameters}{\code{\link[=data.frame]{data.frame()}}}

\item{selectivity_size_info}{\code{\link[=data.frame]{data.frame()}}}

\item{selectivity_age_info}{\code{\link[=data.frame]{data.frame()}}}

\item{selectivity_parameters}{\code{\link[=data.frame]{data.frame()}}}

\item{use_tag_recapture}{\code{\link[=integer]{integer()}}}

\item{use_selectivity_2d}{\code{\link[=integer]{integer()}}}

\item{variance_info}{\code{\link[=data.frame]{data.frame()}}}

\item{lambda_max_phase}{\code{\link[=integer]{integer()}}}

\item{lambda_sd_offset}{\code{\link[=integer]{integer()}}}

\item{lambda_info}{\code{\link[=data.frame]{data.frame()}}}

\item{sd_report_option}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Write Stock Synthesis 3.30 Input Control File
}
\examples{
\dontrun{
# TBD
write_control()
}

}
