% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data.R
\name{write_data}
\alias{write_data}
\title{Write Stock Synthesis 3.30 Data Input File}
\usage{
write_data(
  file = "data.ss",
  year_start,
  year_end,
  n_seasons_per_year,
  n_months_per_season,
  n_subseasons,
  month_spawning,
  n_sexes,
  n_ages,
  n_areas,
  n_fleets,
  fleet_info,
  catch_data,
  index_info,
  index_data,
  n_fleets_discard,
  discard_info,
  discard_data,
  use_mean_body_size,
  length_bin_method,
  length_bin_width,
  length_bin_lower_min,
  length_bin_lower_max,
  use_length_composition,
  length_info,
  n_length_bins,
  length_bins_lower,
  length_data,
  n_age_bins = 0,
  use_mean_size_at_age = 0,
  n_environmental = 0,
  n_size_frequency_methods = 0,
  use_tag_recapture = 0,
  use_morph_composition = 0,
  use_selectivity_data = 0
)
}
\arguments{
\item{file}{\code{\link[=character]{character()}}}

\item{year_start}{\code{\link[=integer]{integer()}}}

\item{year_end}{\code{\link[=integer]{integer()}}}

\item{n_seasons_per_year}{\code{\link[=integer]{integer()}}}

\item{n_months_per_season}{\code{\link[=numeric]{numeric()}}}

\item{n_subseasons}{\code{\link[=integer]{integer()}}}

\item{month_spawning}{\code{\link[=numeric]{numeric()}}}

\item{n_sexes}{\code{\link[=integer]{integer()}}}

\item{n_ages}{\code{\link[=integer]{integer()}}}

\item{n_areas}{\code{\link[=integer]{integer()}}}

\item{n_fleets}{\code{\link[=integer]{integer()}}}

\item{fleet_info}{\code{\link[=data.frame]{data.frame()}}}

\item{catch_data}{\code{\link[=data.frame]{data.frame()}}}

\item{index_info}{\code{\link[=data.frame]{data.frame()}}}

\item{index_data}{\code{\link[=data.frame]{data.frame()}}}

\item{n_fleets_discard}{\code{\link[=integer]{integer()}}}

\item{discard_info}{\code{\link[=data.frame]{data.frame()}}}

\item{discard_data}{\code{\link[=data.frame]{data.frame()}}}

\item{use_mean_body_size}{\code{\link[=integer]{integer()}}}

\item{length_bin_method}{\code{\link[=integer]{integer()}}}

\item{length_bin_width}{\code{\link[=integer]{integer()}}}

\item{length_bin_lower_min}{\code{\link[=integer]{integer()}}}

\item{length_bin_lower_max}{\code{\link[=integer]{integer()}}}

\item{use_length_composition}{\code{\link[=integer]{integer()}}}

\item{length_info}{\code{\link[=data.frame]{data.frame()}}}

\item{n_length_bins}{\code{\link[=integer]{integer()}}}

\item{length_bins_lower}{\code{\link[=integer]{integer()}}}

\item{length_data}{\code{\link[=data.frame]{data.frame()}}}

\item{n_age_bins}{\code{\link[=integer]{integer()}}}

\item{use_mean_size_at_age}{\code{\link[=integer]{integer()}}}

\item{n_environmental}{\code{\link[=integer]{integer()}}}

\item{n_size_frequency_methods}{\code{\link[=integer]{integer()}}}

\item{use_tag_recapture}{\code{\link[=integer]{integer()}}}

\item{use_morph_composition}{\code{\link[=integer]{integer()}}}

\item{use_selectivity_data}{\code{\link[=integer]{integer()}}}
}
\description{
Write Stock Synthesis 3.30 Data Input File
}
\examples{
\dontrun{
write_data()
}
}
