% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-data-frame.R
\name{write_data_frame}
\alias{write_data_frame}
\title{Write Stock Synthesis 3.30 Data Frame}
\usage{
write_data_frame(x, file, append = TRUE)
}
\arguments{
\item{x}{\code{\link[=data.frame]{data.frame()}}}

\item{file}{\code{\link[=character]{character()}}}

\item{append}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=data.frame]{data.frame()}}
}
\description{
Write Stock Synthesis 3.30 Data Frame
}
\examples{
\dontrun{
data_frame <- data.frame(first = 1:10, second = 11:20)
write_data_frame(x = data_frame, file = "temp.ss")
}

}
