% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-forecast.R
\name{write_forecast}
\alias{write_forecast}
\title{Write Stock Synthesis 3.30 Forecast Input File}
\usage{
write_forecast(
  file = "forecast.ss",
  reference_point_option = 1,
  msy_method = 2,
  spr_target = 0.4,
  relative_biomass_target = 0.4,
  reference_point_years = rep(0, 10),
  relative_fishing_basis = 1,
  forecast_option = 0,
  forecast_years = 1,
  fishing_multiplier = 1,
  control_rule_method = 0
)
}
\arguments{
\item{file}{\code{\link[=character]{character()}}}

\item{reference_point_option}{\code{\link[=integer]{integer()}}}

\item{msy_method}{\code{\link[=integer]{integer()}}}

\item{spr_target}{\code{\link[=numeric]{numeric()}}}

\item{relative_biomass_target}{\code{\link[=numeric]{numeric()}}}

\item{reference_point_years}{\code{\link[=integer]{integer()}} \code{\link[=vector]{vector()}}}

\item{relative_fishing_basis}{\code{\link[=integer]{integer()}}}

\item{forecast_option}{\code{\link[=integer]{integer()}}}

\item{forecast_years}{\code{\link[=integer]{integer()}}}

\item{fishing_multiplier}{\code{\link[=numeric]{numeric()}}}

\item{control_rule_method}{\code{\link[=integer]{integer()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Write Stock Synthesis 3.30 Forecast Input File
}
\examples{
\dontrun{
write_forecast()
}

}
