% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-starter.R
\name{write_starter}
\alias{write_starter}
\title{Write Stock Synthesis 3.30 Starter Input File}
\usage{
write_starter(
  file = "starter.ss",
  file_name_data = "data.ss",
  file_name_control = "control.ss",
  use_ss3_dot_par = 0,
  progress_detail = 0,
  age_report_option = 1,
  custom_report_option = 0,
  parameter_trace = 0,
  cumulative_report_option = 2,
  use_full_priors = 0,
  use_soft_bounds = 1,
  data_output_option = 1,
  phase_maximum = 10,
  mcmc_burn = 250,
  mcmc_thin = 5,
  jitter_option = 0,
  sd_report_year_min = -1,
  sd_report_year_max = -1,
  sd_report_years_extra = 0,
  convergence_value = 1e-04,
  year_retrospective = 0,
  age_summary_option = 1,
  depletion_basis = 1,
  depletion_denom = 1,
  spr_basis = 0,
  fishing_units = 1,
  fishing_basis = 0,
  mcmc_detail = 2,
  alk_tolerance = 0,
  boot_seed = 0,
  check_value = "3.30"
)
}
\arguments{
\item{file}{\code{\link[=character]{character()}}}

\item{file_name_data}{\code{\link[=character]{character()}}}

\item{file_name_control}{\code{\link[=character]{character()}}}

\item{use_ss3_dot_par}{\code{\link[=integer]{integer()}}}

\item{progress_detail}{\code{\link[=integer]{integer()}}}

\item{age_report_option}{\code{\link[=integer]{integer()}}}

\item{custom_report_option}{\code{\link[=integer]{integer()}}}

\item{parameter_trace}{\code{\link[=integer]{integer()}}}

\item{cumulative_report_option}{\code{\link[=integer]{integer()}}}

\item{use_full_priors}{\code{\link[=integer]{integer()}}}

\item{use_soft_bounds}{\code{\link[=integer]{integer()}}}

\item{data_output_option}{\code{\link[=integer]{integer()}}}

\item{phase_maximum}{\code{\link[=integer]{integer()}}}

\item{mcmc_burn}{\code{\link[=integer]{integer()}}}

\item{mcmc_thin}{\code{\link[=integer]{integer()}}}

\item{jitter_option}{\code{\link[=numeric]{numeric()}}}

\item{sd_report_year_min}{\code{\link[=integer]{integer()}}}

\item{sd_report_year_max}{\code{\link[=integer]{integer()}}}

\item{sd_report_years_extra}{\code{\link[=integer]{integer()}}}

\item{convergence_value}{\code{\link[=numeric]{numeric()}}}

\item{year_retrospective}{\code{\link[=integer]{integer()}}}

\item{age_summary_option}{\code{\link[=integer]{integer()}}}

\item{depletion_basis}{\code{\link[=integer]{integer()}}}

\item{depletion_denom}{\code{\link[=numeric]{numeric()}}}

\item{spr_basis}{\code{\link[=integer]{integer()}}}

\item{fishing_units}{\code{\link[=integer]{integer()}}}

\item{fishing_basis}{\code{\link[=integer]{integer()}}}

\item{mcmc_detail}{\code{\link[=integer]{integer()}}}

\item{alk_tolerance}{\code{\link[=integer]{integer()}}}

\item{boot_seed}{\code{\link[=numeric]{numeric()}}}

\item{check_value}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Write Stock Synthesis 3.30 Starter Input File
}
\examples{
\dontrun{
write_starter()
}

}
