% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-terminal-line.R
\name{write_terminal_line}
\alias{write_terminal_line}
\title{Write Stock Synthesis 3.30 Data Frame Terminal Line}
\usage{
write_terminal_line(
  n,
  file,
  value = -9999,
  name = "terminal line",
  append = TRUE
)
}
\arguments{
\item{n}{\code{\link[=integer]{integer()}}}

\item{file}{\code{\link[=character]{character()}}}

\item{value}{\code{\link[=numeric]{numeric()}}}

\item{name}{\code{\link[=character]{character()}}}

\item{append}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=character]{character()}}
}
\description{
Write Stock Synthesis 3.30 Data Frame Terminal Line
}
\examples{
\dontrun{
write_terminal_line(n = 10, file = "temp.ss")
}

}
